slackに送信するロジックを共通化
