+++ Dockerfile
@@ -4,5 +4,6 @@ WORKDIR /app
 COPY package.json .
 COPY index.js .
 COPY preStop.js .
+COPY slack.js .
 
 CMD [ "node", "/app/index.js" ]
