+++ index.js
@@ -1,4 +1,4 @@
-const https = require('https');
+const sendSlack = require('./slack').sendSlack;
 
 writeLog();
 const timer = setInterval(() => {
@@ -27,23 +27,4 @@ function endProcess(signal = 'SIGTERM') {
 		sendSlack(`${(new Date()).toLocaleString()} still alive(after ${i} seconds).`);
 		i++;
 	}, 1000);
-}
-
-function sendSlack(text) {
-	const req = https.request('https://slack.com/api/chat.postMessage', {
-		method: 'POST',
-		headers: {
-			Authorization: `Bearer ${process.env['SLACK_TOKEN']}`,
-			'Content-Type': 'application/json; charset=UTF-8',
-		}
-	}, (res) => {
-		res.setEncoding('utf8');
-		res.on('data', (chunk) => {
-			console.log(chunk);
-		});
-		res.on('end', () => {});
-	});
-
-	req.write(JSON.stringify({ channel: 'post-message', text }));
-	req.end();
 }
